Finalized breakout of utility functions and updated references.
