+++ ReQuest/cogs/admin.py
@@ -6,7 +6,7 @@ from discord.ext.commands import Cog
 from ReQuest.ui.admin import views
 from ReQuest.utilities.checks import is_owner
 from ReQuest.utilities.localizer import t, DEFAULT_LOCALE, resolve_locale, set_locale_context
-from ReQuest.utilities.supportFunctions import log_exception
+from ReQuest.utilities.exceptions import log_exception
 
 
 class Admin(Cog):
