+++ ReQuest/cogs/shop.py
@@ -5,7 +5,9 @@ from discord.ext.commands import Cog
 from ReQuest.ui.shop import views
 from ReQuest.utilities.constants import CommonFields, ShopFields, DatabaseCollections
 from ReQuest.utilities.localizer import resolve_locale, set_locale_context, t
-from ReQuest.utilities.supportFunctions import log_exception, get_cached_data, UserFeedbackError, setup_view
+from ReQuest.utilities.db_cache import get_cached_data
+from ReQuest.utilities.exceptions import UserFeedbackError, log_exception
+from ReQuest.utilities.discord_utils import setup_view
 
 
 class Shop(Cog):
