+++ ReQuest/ui/config/buttons.py
@@ -14,17 +14,11 @@ from ReQuest.utilities.constants import (
     ConfigFields, ShopFields, CommonFields, RoleplayFields, CurrencyFields, DatabaseCollections
 )
 from ReQuest.utilities.localizer import t, DEFAULT_LOCALE, resolve_locale
-from ReQuest.utilities.supportFunctions import (
-    log_exception,
-    setup_view,
-    get_cached_data,
-    delete_cached_data,
-    update_cached_data,
-    get_xp_config,
-    remove_item_stock_limit,
-    encode_mongo_key,
-    format_currency_amount
-)
+from ReQuest.utilities.currency import format_currency_amount
+from ReQuest.utilities.db_cache import get_cached_data, delete_cached_data, update_cached_data, encode_mongo_key, get_xp_config
+from ReQuest.utilities.exceptions import log_exception
+from ReQuest.utilities.discord_utils import setup_view
+from ReQuest.utilities.shop import remove_item_stock_limit
 
 logger = logging.getLogger(__name__)
