+++ ReQuest/ui/shop/buttons.py
@@ -5,15 +5,9 @@ from discord.ui import Button
 from ReQuest.ui.shop import modals
 from ReQuest.utilities.constants import CharacterFields, ShopFields, CommonFields, CartFields, DatabaseCollections
 from ReQuest.utilities.localizer import t, DEFAULT_LOCALE
-from ReQuest.utilities.supportFunctions import (
-    log_exception,
-    get_cached_data,
-    UserFeedbackError,
-    clear_cart_and_release_stock,
-    get_shop_stock,
-    get_cart,
-    get_item_stock
-)
+from ReQuest.utilities.db_cache import get_cached_data
+from ReQuest.utilities.exceptions import UserFeedbackError, log_exception
+from ReQuest.utilities.shop import clear_cart_and_release_stock, get_shop_stock, get_cart, get_item_stock
 
 
 class ShopItemButton(Button):
