+++ ReQuest/utilities/db_cache.py
@@ -1,16 +1,56 @@
 import json
 import logging
 
-from ReQuest.utilities.constants import CommonFields
+from ReQuest.utilities.constants import CommonFields, ConfigFields, DatabaseCollections
 from ReQuest.utilities.exceptions import log_exception
 
 logger = logging.getLogger(__name__)
 
-__all__ = ['build_cache_key', 'get_cached_data', 'update_cached_data', 'replace_cached_data', 'delete_cached_data']
+__all__ = [
+    'build_cache_key', 'get_cached_data', 'update_cached_data', 'replace_cached_data', 'delete_cached_data',
+    'encode_mongo_key', 'decode_mongo_key', 'get_xp_config',
+]
 
 
 def build_cache_key(database_name, identifier, collection_name):
     return f'{database_name}:{identifier}:{collection_name}'
+
+
+def encode_mongo_key(key: str) -> str:
+    """
+    Encodes a string for safe use as a MongoDB field name.
+
+    MongoDB field names cannot contain:
+    - '.' (dot) - interpreted as nested document path
+    - '$' (dollar) - reserved for operators
+    - null characters - not allowed in field names
+
+    Uses URL-encoding style. The '%' character is encoded first to ensure reversibility.
+    """
+    if not key:
+        return key
+    # Order matters: encode '%' first to avoid double-encoding
+    result = key.replace('%', '%25')
+    result = result.replace('.', '%2E')
+    result = result.replace('$', '%24')
+    # Strip null characters (invalid in MongoDB field names and have no display value)
+    result = result.replace('\x00', '')
+    return result
+
+
+def decode_mongo_key(key: str) -> str:
+    """
+    Decodes a MongoDB field name back to its original form.
+
+    Reverses the encoding applied by encode_mongo_key().
+    """
+    if not key:
+        return key
+    # Order matters: decode '%' last to avoid incorrect decoding
+    result = key.replace('%2E', '.')
+    result = result.replace('%24', '$')
+    result = result.replace('%25', '%')
+    return result
 
 
 async def get_cached_data(bot, mongo_database, collection_name, query, is_single=True, cache_id=None):
@@ -179,3 +219,28 @@ async def delete_cached_data(bot, mongo_database, collection_name, search_filter
         await bot.rdb.delete(cache_key)
     except Exception as e:
         logger.error(f"Redis delete failed: {e}")
+
+
+async def get_xp_config(bot, guild_id) -> bool:
+    """
+    Retrieves the XP configuration for a guild.
+
+    :param bot: The Discord bot instance
+    :param guild_id: The Discord guild id
+
+    :return: True if XP is enabled, False if XP is disabled
+    """
+    try:
+        query = await get_cached_data(
+            bot=bot,
+            mongo_database=bot.gdb,
+            collection_name=DatabaseCollections.PLAYER_EXPERIENCE,
+            query={CommonFields.ID: guild_id}
+        )
+        if query is None:
+            return True  # Default to XP enabled if no config found
+        return query.get(ConfigFields.PLAYER_EXPERIENCE, True)
+    except Exception as e:
+        logger.error(f"Error retrieving XP config: {e}")
+        await log_exception(e)
+        return True  # Default to XP enabled on error
