+++ ReQuest/utilities/discord_utils.py
@@ -1,16 +1,21 @@
+import inspect
 import logging
 import re
 
 import discord
 
+from ReQuest.utilities.exceptions import UserFeedbackError
+
 logger = logging.getLogger(__name__)
 
 __all__ = [
     'attempt_delete',
+    'check_role_hierarchy',
     'strip_id',
     'truncate_text',
     'escape_markdown',
     'get_guild_member',
+    'setup_view',
 ]
 
 
@@ -24,6 +29,19 @@ async def attempt_delete(message: discord.Message | discord.PartialMessage):
         logger.error(f'HTTPException while deleting message: {e}')
     except Exception as e:
         logger.error(f'Unexpected error while deleting message: {e}')
+
+
+def check_role_hierarchy(guild: discord.Guild, role: discord.Role):
+    """Raises UserFeedbackError if the bot cannot manage the given role due to hierarchy."""
+    from ReQuest.utilities.localizer import t, DEFAULT_LOCALE
+    bot_top_role = guild.me.top_role
+    if role >= bot_top_role:
+        raise UserFeedbackError(
+            t(DEFAULT_LOCALE, 'gm-error-role-hierarchy', roleName=role.name, roleId=str(role.id)),
+            message_id='gm-error-role-hierarchy',
+            roleName=role.name,
+            roleId=str(role.id)
+        )
 
 
 def strip_id(mention: str) -> int:
@@ -66,6 +84,34 @@ def escape_markdown(text: str) -> str:
     for char in ('*', '_', '~', '`', '|', '>', '[', ']', '(', ')'):
         text = text.replace(char, f'\\{char}')
     return text
+
+
+async def setup_view(view, interaction: discord.Interaction):
+    """
+    Dynamically sets up a view by inspecting its setup method for required parameters.
+    Resolves and propagates the user's locale to the view before calling setup().
+    """
+    from ReQuest.utilities.localizer import resolve_locale, set_locale_context
+
+    locale = await resolve_locale(interaction)
+    set_locale_context(locale)
+    view.locale = locale
+
+    setup_function = view.setup
+    sig = inspect.signature(setup_function)
+    params = sig.parameters
+
+    kwargs = {}
+    if 'bot' in params:
+        kwargs['bot'] = interaction.client
+    if 'user' in params:
+        kwargs['user'] = interaction.user
+    if 'guild' in params:
+        kwargs['guild'] = interaction.guild
+    if 'interaction' in params:
+        kwargs['interaction'] = interaction
+
+    await setup_function(**kwargs)
 
 
 async def get_guild_member(guild: discord.Guild, user_id: int) -> discord.Member | None:
